requirements.txt
streamlit
sentence-transformers
requests
googlesearch-python
syncedlyrics